+++ sphinx/source/changelog.rst
@@ -5,6 +5,8 @@ Changelog
 -------
 * When Escape is pressed to quit the game, a confirmation screen is displayed
   (unless the ``ConfirmQuit`` parameter in the :ref:`gameConfig` section is 0)
+* Added support for appending content to existing ini file sections by adding a
+  '+' suffix to the section name (e.g. `[SkoolLocations+]`)
 * Added the ``--force`` command line option (to overwrite existing images, ini
   files and sound files)
 * Added the ``--sample-rate`` command line option (to set the sample rate of